Refactor orchestration helpers into focused modules
